+++ energiminer/solution.h
@@ -38,7 +38,7 @@ public:
     Solution()
     {}
 
-    Solution(Work work, uint32_t nonce, const uint256& mixhash)
+    Solution(Work work, uint64_t nonce, const uint256& mixhash)
         : m_nonce(nonce)
         , m_mixhash(mixhash)
         , m_work(work)
@@ -59,7 +59,7 @@ public:
     }
 
 public:
-    uint32_t m_nonce;
+    uint64_t m_nonce;
     uint256 m_mixhash;
 
 private:
